backend/pyproject.toml
[tool.poetry]
name = "nt-center-invest"
version = "0.1.0"
description = ""
authors = ["CoSm1C AvEnGeR <zhora.zhilin.06@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.2"}
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
motor = "^3.4.0"
beanie = "^1.25.0"
fastapi-users = {extras = ["beanie"], version = "^13.0.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
